Refactor hex formatted
