+++ crates/client/src/client/types/primitives.rs
@@ -1,72 +1,25 @@
-use crate::client::schema::ConversionError;
-
-use serde::{
-    de::Error,
-    Deserialize,
-    Deserializer,
-    Serialize,
-    Serializer,
+use crate::client::{
+    HexFormatError,
+    HexFormatted,
 };
+
 use std::{
     fmt::{
         Debug,
         Display,
         Formatter,
-        LowerHex,
     },
     str::FromStr,
 };
 use tai64::Tai64;
 
-#[derive(Debug, Clone, Default)]
-pub struct HexFormatted<T: Debug + Clone + Default>(pub T);
-
-impl<T: LowerHex + Debug + Clone + Default> Serialize for HexFormatted<T> {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        serializer.serialize_str(format!("{:#x}", self.0).as_str())
-    }
-}
-
-impl<'de, T: FromStr<Err = E> + Debug + Clone + Default, E: Display> Deserialize<'de>
-    for HexFormatted<T>
-{
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let s: String = Deserialize::deserialize(deserializer)?;
-        T::from_str(s.as_str()).map_err(D::Error::custom).map(Self)
-    }
-}
-
-impl<T: FromStr<Err = E> + Debug + Clone + Default, E: Display> FromStr
-    for HexFormatted<T>
-{
-    type Err = ConversionError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        T::from_str(s)
-            .map_err(|e| ConversionError::HexError(format!("{e}")))
-            .map(Self)
-    }
-}
-
-impl<T: LowerHex + Debug + Clone + Default> Display for HexFormatted<T> {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(f, "0x{:x}", self.0)
-    }
-}
-
-macro_rules! fuel_type_scalar {
+macro_rules! client_type_scalar {
     ($id:ident, $base_id:ident) => {
         #[derive(Debug, Clone, Default)]
         pub struct $id(pub HexFormatted<base::$base_id>);
 
         impl FromStr for $id {
-            type Err = ConversionError;
+            type Err = HexFormatError;
 
             fn from_str(s: &str) -> Result<Self, Self::Err> {
                 let b = HexFormatted::<base::$base_id>::from_str(s)?;
@@ -202,18 +155,18 @@ pub mod base {
     }
 }
 
-fuel_type_scalar!(Address, Bytes32);
-fuel_type_scalar!(AssetId, Bytes32);
-fuel_type_scalar!(BlockId, Bytes32);
-fuel_type_scalar!(ContractId, Bytes32);
-fuel_type_scalar!(MerkleRoot, Bytes32);
-fuel_type_scalar!(Message, Bytes32);
-fuel_type_scalar!(MessageId, Bytes32);
-fuel_type_scalar!(Nonce, Bytes32);
-fuel_type_scalar!(PublicKey, Bytes64);
-fuel_type_scalar!(Salt, Bytes32);
-fuel_type_scalar!(Signature, Bytes64);
-fuel_type_scalar!(TransactionId, Bytes32);
+client_type_scalar!(Address, Bytes32);
+client_type_scalar!(AssetId, Bytes32);
+client_type_scalar!(BlockId, Bytes32);
+client_type_scalar!(ContractId, Bytes32);
+client_type_scalar!(MerkleRoot, Bytes32);
+client_type_scalar!(Message, Bytes32);
+client_type_scalar!(MessageId, Bytes32);
+client_type_scalar!(Nonce, Bytes32);
+client_type_scalar!(PublicKey, Bytes64);
+client_type_scalar!(Salt, Bytes32);
+client_type_scalar!(Signature, Bytes64);
+client_type_scalar!(TransactionId, Bytes32);
 
 #[derive(Debug)]
 pub struct Tai64Timestamp(pub Tai64);
